+++ ETicket/ETicketWebAPI/Models/TicketVerification/VerifyTicketRequest.cs
@@ -5,7 +5,7 @@ namespace ETicket.WebAPI.Models.TicketVerification
     public class VerifyTicketRequest
     {
         public Guid TicketId { get; set; }
-        public long TransportId { get; set; }
+        public int TransportId { get; set; }
         public float Longitude { get; set; }
         public float Latitude { get; set; }
     }
